move vertex client tests to a separate file
